implemented MKN lhs@rhs
